requirements.txt
pytest>=8.0.0
allure-pytest>=2.13.0
pyyaml>=6.0
requests>=2.31.0
